The game can be played now by selecting START GAME.
The game now enters in menu mode and if "START GAME" button is pressed
then the game enters in game play mode. If the ESC key is pressed then
it returns back to the menu mode. In menu mode the game exits with
pressing the "EXIT GAME" button.
The setting the different modes is handled by the event holder but not
by the game class anymore.
Some minor code refactories.
Some unnecessary code regarding loading fonts on unnecessary place
deleted.