Timer level introduced.
Game over is triggered if Zero fail to reach to its home before the time ends. The timer displays the remaining time at the top of the screen in green unless the remaining time is less than 15 seconds - then it is in red. In the last 10 seconds a sound is played each second;
Minor refactorings