FIX: pressing eSC key game & menu behavior
Fix if by pressing the ESC key during the possible game states. Previously, if pressing the ESC key outside the game mode then the game mode will be bring out. But this means it to be bring during the Start screen, the final screen and in any sub-menu.
Now this fixes the ESC pressing key behavior: if not in game mode and if in resume game menu - bring the game mode; if in any of the submenus - bring the previous menu (currently Main or Resume Game).

Minor fix: enum naming

Update: ToDo list