/Not working yet/ Add: the Load Game menu
The main shows the Load Game menu instead of the Save Game menu.
The resume game menus shows the Load Game menu above the Save Game.
The Load Game menu shows currently saved games slots. But it does just
that - just showing them. If trying to load a game from a slot then it
will result in just going back to the game